update migration file

+++ packages/shared/src/migrations/1714987069557-createSocketIoAttachmentsTable.js
@@ -1,10 +1,13 @@
 /** @typedef {import('sequelize').QueryInterface} QueryInterface */
 import { DataTypes, Sequelize } from 'sequelize';
+import config from 'config';
 
 /**
  * @param {QueryInterface} query
  */
 export async function up(query) {
+  if (config.serverFacilityId) return;
+
   await query.createTable('socket_io_attachments', {
     id: {
       type: DataTypes.UUID,
@@ -26,5 +29,7 @@ export async function up(query) {
  * @param {QueryInterface} query
  */
 export async function down(query) {
+  if (config.serverFacilityId) return;
+
   await query.dropTable('socket_io_attachments');
 }
